#XGAL_37

+++ app/Console/Commands/Jav/JavDownloads.php
@@ -11,12 +11,14 @@ namespace App\Console\Commands\Jav;
 
 use App\Console\BaseCommand;
 use App\Crawlers\Crawler\Onejav;
+use App\Facades\UserActivity;
+use App\Models\JavDownload;
 
 /**
  * Process download pending JAV
  * @package App\Console\Commands
  */
-final class JavDownload extends BaseCommand
+final class JavDownloads extends BaseCommand
 {
     /**
      * The name and signature of the console command.
@@ -34,7 +36,7 @@ final class JavDownload extends BaseCommand
 
     protected function download()
     {
-        $downloads = \App\Models\JavDownload::all();
+        $downloads = JavDownload::all();
         if ($downloads->isEmpty()) {
             return true;
         }
@@ -43,11 +45,19 @@ final class JavDownload extends BaseCommand
         $downloads->each(function ($download) {
             $this->progressBarSetInfo($download->item_number);
             $item = $download->downloads()->first();
+
+            if (!$item) {
+                return;
+            }
+
             $crawler = app(Onejav::class);
+            // Check again to get updated torrent link
             $item = $crawler->getItems($item->url)->first();
             $crawler->getClient()->download($item->torrent, 'onejav');
             $this->progressBarSetStatus('FINISHED');
             $download->forceDelete();
+
+            UserActivity::notify('%s %s video ' . $item->title, null, 'download');
         });
 
         return true;
